refactor: split processor
