+++ program/rust/src/processor.rs
@@ -5,23 +5,6 @@ use {
             load_command_header_checked,
             OracleCommand,
         },
-        rust_oracle::{
-            add_mapping,
-            add_price,
-            add_product,
-            add_publisher,
-            del_price,
-            del_product,
-            del_publisher,
-            init_mapping,
-            init_price,
-            resize_price_account,
-            set_min_pub,
-            upd_permissions,
-            upd_price,
-            upd_price_no_fail_on_error,
-            upd_product,
-        },
     },
     solana_program::{
         entrypoint::ProgramResult,
@@ -30,34 +13,68 @@ use {
     },
 };
 
+mod add_mapping;
+mod add_price;
+mod add_product;
+mod add_publisher;
+mod del_price;
+mod del_product;
+mod del_publisher;
+mod init_mapping;
+mod init_price;
+mod resize_price_account;
+mod set_min_pub;
+mod upd_permissions;
+mod upd_price;
+mod upd_product;
+
+pub use {
+    add_mapping::add_mapping,
+    add_price::add_price,
+    add_product::add_product,
+    add_publisher::add_publisher,
+    del_price::del_price,
+    del_product::del_product,
+    del_publisher::del_publisher,
+    init_mapping::init_mapping,
+    init_price::init_price,
+    resize_price_account::resize_price_account,
+    set_min_pub::set_min_pub,
+    upd_permissions::upd_permissions,
+    upd_price::{
+        c_upd_aggregate,
+        upd_price,
+        upd_price_no_fail_on_error,
+    },
+    upd_product::upd_product,
+};
+
 /// Dispatch to the right instruction in the oracle.
 pub fn process_instruction(
     program_id: &Pubkey,
     accounts: &[AccountInfo],
     instruction_data: &[u8],
 ) -> ProgramResult {
+    use OracleCommand::*;
+
     match load_command_header_checked(instruction_data)? {
-        OracleCommand::InitMapping => init_mapping(program_id, accounts, instruction_data),
-        OracleCommand::AddMapping => add_mapping(program_id, accounts, instruction_data),
-        OracleCommand::AddProduct => add_product(program_id, accounts, instruction_data),
-        OracleCommand::UpdProduct => upd_product(program_id, accounts, instruction_data),
-        OracleCommand::AddPrice => add_price(program_id, accounts, instruction_data),
-        OracleCommand::AddPublisher => add_publisher(program_id, accounts, instruction_data),
-        OracleCommand::DelPublisher => del_publisher(program_id, accounts, instruction_data),
-        OracleCommand::UpdPrice => upd_price(program_id, accounts, instruction_data),
-        OracleCommand::AggPrice => upd_price(program_id, accounts, instruction_data),
-        OracleCommand::InitPrice => init_price(program_id, accounts, instruction_data),
-        OracleCommand::InitTest => Err(OracleError::UnrecognizedInstruction.into()),
-        OracleCommand::UpdTest => Err(OracleError::UnrecognizedInstruction.into()),
-        OracleCommand::SetMinPub => set_min_pub(program_id, accounts, instruction_data),
-        OracleCommand::UpdPriceNoFailOnError => {
-            upd_price_no_fail_on_error(program_id, accounts, instruction_data)
-        }
-        OracleCommand::ResizePriceAccount => {
-            resize_price_account(program_id, accounts, instruction_data)
-        }
-        OracleCommand::DelPrice => del_price(program_id, accounts, instruction_data),
-        OracleCommand::DelProduct => del_product(program_id, accounts, instruction_data),
-        OracleCommand::UpdPermissions => upd_permissions(program_id, accounts, instruction_data),
+        InitMapping => init_mapping(program_id, accounts, instruction_data),
+        AddMapping => add_mapping(program_id, accounts, instruction_data),
+        AddProduct => add_product(program_id, accounts, instruction_data),
+        UpdProduct => upd_product(program_id, accounts, instruction_data),
+        AddPrice => add_price(program_id, accounts, instruction_data),
+        AddPublisher => add_publisher(program_id, accounts, instruction_data),
+        DelPublisher => del_publisher(program_id, accounts, instruction_data),
+        UpdPrice => upd_price(program_id, accounts, instruction_data),
+        AggPrice => upd_price(program_id, accounts, instruction_data),
+        InitPrice => init_price(program_id, accounts, instruction_data),
+        InitTest => Err(OracleError::UnrecognizedInstruction.into()),
+        UpdTest => Err(OracleError::UnrecognizedInstruction.into()),
+        SetMinPub => set_min_pub(program_id, accounts, instruction_data),
+        UpdPriceNoFailOnError => upd_price_no_fail_on_error(program_id, accounts, instruction_data),
+        ResizePriceAccount => resize_price_account(program_id, accounts, instruction_data),
+        DelPrice => del_price(program_id, accounts, instruction_data),
+        DelProduct => del_product(program_id, accounts, instruction_data),
+        UpdPermissions => upd_permissions(program_id, accounts, instruction_data),
     }
 }
